Add Old Maui as a test of MD description fields
This is the output that Netlify CMS generates in order to create a
markdown field inside the frontmatter. It's here in the case that we
want formatting (for example images) in our song description body to
keep the actual song text minimalist.

Eleventy seems not to barf at this format, so we can use it for now

Note that we seem to need to use:
templateEngineOverride: njk,md

in any place we want proper Markdown for these fields, see
https://github.com/11ty/eleventy/issues/236