Add songs with song numbers
This commit includes a couple of steps
1. Add a basic page type for songs, and a songlist page+component
2. Include a 'custom collection' for numbered songs, injecting the
   number at site generation time.

This is a first pass approach to numbering - we may well not want it
to change each time we generate the site (imagine we've printed one
run of books already!) so something to think about. songNumber could
easily be re-injected back into the MD as a 1-stage process.