Add 'song number' redirects
The point of having numbered song entries is that we can redirect
easily to a number that someone calls out.

This uses pagination and some hackery to create a set of pages at
URLs like /n/X were X is the song number, that redirect you to the
correct page. I'm leaving the correct pages in for SEO and general
URL sanity, and redirecting on these numbered pages.