Update markdownit and minifier options, adjust songs
The boilerplate I'd used didn't actually include the required call
to set the config for markdown-it so it the 'breaks:true' option
was not being applied. This fixes that

It also removes 'collapse whitespace' on the minifier, which is
useful as we're working on the songs

Finally I've adjusted just Salt Liberty to test the fact we don't
need to escape line endings now, and Old Maui to test removing
escaped apostrophes (also fine).