Add nav bar using Bulma, first pass
This includes bulma.css and adds a functioning nav bar. I've moved
some things down to a second level nav, and created a button (not
yet working) to jump to a song number.

This is on a branch as I'm not sure bulma.css is really what we
want and I think potentially we'd be better going for something more
normal like bootstrap, not least because there are more examples to
help me learn stuff as I go